fix: Correct the LRD targeter display (CanOfWorms)
The fragmentation targeter's display wasn't properly accounting for
intervening walls and statues and indicated additional cells as being
affected that were not in the final explosion. This was because the
targeter was running the explosion map determination code with the
incorrect arguments and beam flavor. The actual LRD explosion was
correct, since that's run later, and in the correct way.

This commit corrects the targeter logic so it agrees with the actual
explosion.

(cherry picked from commit 6621cef402bd1940e907d5b0c04ec30777f15900)