fix: improve handling of numpad keycodes on webtiles
Modern browsers can differentiate numpad keycodes from their regular
versions quite straightforwardly, but this information wasn't being
handled well by the javascript interface. This was manageable on devices
that have numlock, because with numlock off, the keycodes for the
secondary keycodes are usually fine for movement, and the number keys
are available for macros. But on devices without numlock (like all macs
for many years), this was pretty annoying if you use the numpad for
movement -- there was no way to use the numrow number keys separately,
and rebinding them rebound the number pad keys. Moreover, these keys
were generally handled better on local console/tiles.

This commit attempts to remedy this situation along with some other
tweaks.
* key handling in the webtiles client now has a versioned component that
  can override unversioned key handling.
* Support for legacy F keycodes (see b0746d5dc15c) is dropped going
  forward. This will break some existing macros that used the (wrong)
  values in the range -1011 to -1014, these should be easy enough to fix
  up but basically the correct values are -265 to -268.
* The client uses w3c `code` values[1] to detect numpad keys in a way that
  works across browsers and maps them to (a version of) the internal
  keycodes that local console was already using for them. This means
  that default movement bindings etc continue to work for them.
* I did some standardization and cleanup of these keys for
  non-local-tiles builds, and their names will show in the macro editor.

[1] https://www.w3.org/TR/uievents-code/#code-value-tables

Caveats:
* The underlying curses code remains quite shaky, and numpad = is
  currently not handled at all.
* The versioned keymapping API is non-existent; and in the future we may
  want to consider modernizing this to check `code`.
* Probably more variation across devices / builds than is accounted for
  here...