fix: only normalize numpad number keycodes
The key conditional in 7b1be97da7 was converting too many cases; for
example it was possible for `code` to be NumpadX with arrows/home/etc.
This also should fix shift handling for a number of cases, depending on
what the keyboard actually sends.

* report 1 (Ham): shift-5 sends clear. This should now be fixed for this
  key configuration, as I think it was being prevented by keycode
  overwriting. It's worth noting that this is a pretty non-standard
  arrangement, and since we have no specific translation for shift-5, it
  probably just sends numeric 5 on many keyboards (as it does on mine).
  I'll await more reports to see if there should be some standardization
  for shift+5. Shift+0 is in a similar state.
* report 2 (dinky): `code` values lined up with number keys, but `key`
  and `keyCode` are arrow/home/etc. keys. Interestingly, no keypress
  event at all. Again, fixed by reining in keycode replacement.

We'll see what other exciting keyboard situations the world provides
next...