fix: minimize impact of numpad changes on old versions
This commit should restore the exact (sometimes broken, but mostly
semi-working) behavior of the numpad on pre-0.27 versions. I've seen
reports of issues on e.g. 0.26 where it's a bit tricky to identify the
exact cause, and so this is to try to eliminate them. Possibly, some of
this can be later removed, e.g. this restores a browser-specific chrome
tweak that doesn't line up with my testing of current chrome in the wild
(that's why I originally removed it).