fix: map keycodes for numpad number keys only
This is an extremely limited version of 7b1be97da753 for the sake of not
pushing untested code near a release. This commit maps browser keycodes
to the internal keycodes for numpad number keys only. These keycodes
already had bindings in cmd-keys.h, so they behave right without any
change on the client side. The rest of 7b1be97da753 is bells and
whistles to make them look good in the macro menu, and handling for the
extra keys like the operator keys, which behave just as they did before.

(This is relying on unversioned key handling code that is already in
trunk.)