New spell: Explosive Dismissal (L6 Conj)
A combination of the old Aura of Abjuration and Forceful Dismissal spells,
Explosive Dismissal causes all hostile summons in sight to explode, dealing
damage to themselves and everything around them. It's intended to be a spell
that makes players feel smart for turning enemies' strengths against them,
in the same way that Yara's Unravelling does for enemy buffs.

It's specifically not a Summoning spell because big explosions everywhere
play badly with allies, and the Ignition trick of not damaging allies felt
weird. I considered making it Cj/Hex or Cj/Tmut, but just Cj seemed fine.

A variation on this idea would be a spell that deals damage to summons &
their summoners, rather than messing with explosions to *probably* hit the
summoner. That seemed less fun, though.

It's possible that this ends up playing too much like Ignition in practice.
I'm not sure! It seemed interesting enough to implement.