fix: temporarily disable new keymappings except 0-9
Because the previous commit fixes this in code that usually takes a
while to make its way into people's caches, I'm just temporarily
bringing trunk's keypad and fn key behavior in line with 0.27, in hopes
of getting the persistance bug to go away quicker.

This commit partially readds some of the legacy handling for function
keycodes in the -1000 range, so that their names aren't shown
confusingly for anyone actually using trunk while this commit is in
place.

I expect to revert this commit in the near future.