fix: extend aa6b198b to keydown handling (Spaniq, wheals)
This will be more impactful once full numpad key handling is back on,
but for now it is also impactful on keyboards that send a
high/numpad-specialized keycode for some numpad operator keys together
with a regular keypress (I can reproduce this on mine).

I think this override (and the one in aa6b198b) could be made better by
checking whether the menu is in key entry mode specifically, it doesn't
need to happen for the entire menu. However, I don't want to mess with
this too much just now.