fix: handle clicking on non-chr hotkey items correctly
When input is sent via `data`, it is passed through `chr()` in the
python code. But, the value stored here is a keycode and may be negative
for special keys, so this was causing an exception in python. This comes
up specifically when clicking on macro menu entries for some special keys,
and pretty much nowhere else (the main menu uses esc, but this is still
a positive keycode). On the user side, it's relatively harmless, the
symptom is just that clicking doesn't work, so backporting to stable is
not urgent. It just produces log messages.

Two changes around this:
* for arrow-selection menus, don't use the hotkey at all, force calling
  on_singleselect. (Hopefully in the future this distinction is better
  cleaned up.)
* send hotkeys via `keycode` rather than `data`, so that this code is
  more robust for the future. Should have no impact on current menus...