fix: speed up explosion calculations (gammafunk)
The check for whether a cell had already been seen with count value <= n
was not properly taking into account the additive factor cadd that is
accumulated on recursion, and so was redoing a lot of calculations to no
extra effect. This led to big slowdowns on e.g. firestorm targeters in
open areas. Or at least, that's what I *think* was happening, this
algorithm is pretty complicated. But this change speeds things up quite
a bit, and doesn't change explosion shapes as far as I can tell.

N.b. this function is still called many times during targeting, but
without noticeable slowdown after this change, in my profiling.