Give Maxwell's LOS range
Range was limited to 5 to avoid risk-free murder of statues, but,
as a level 8 spell, it's probably fine to murder statues freely.
It's simpler to reason about the spell if it can hit anywhere in
LOS - players already get confused about the delayed targeting, so
we don't need to add more space for ambiguity. (If it did miss due
to being outside range, it'd be quite unclear what had happened.)