feat: Don't use trees in a Swamp transporter vault
The vault gammafunk_worm_tunnel uses a border of trees to help it blend
in with the surrounding level. It's trivial to burn some of these trees
with an item as common as scroll of immolation and collect the loot in a
way that bypasses all threats in the vault. Also, the vault's size has
been greatly reduced, so it doesn't stand out in a drastically weird
way. This commit restores the stone walls when it places in Swamp, so it
uses the same wall type in all places it generates.

(cherry picked from commit 77753f3aa732dbfa16f54e9dd6fe43829cab851e)