fix: handle multiple SDL key events for numpad `.`
For some reason, SDL is sending both . and Del for numpad . on some
keyboards, even when . isn't delete. This commit handles that case by
suppressing the . when there's a double event, and providing a default
mapping of delete to CMD_WAIT. This should be fine because it's
previously unbound, but I've set this to local tiles only; it seems
buggy that SDL is sending this keycode at all (at least on my keyboard),
so this may be a temporary fix if I can figure out what is going on
here. SDL numpad handling generally needs some attention.

For background on the key handling issues, see af9c1943b0bc.

Reported by discord user Dank Memer.