fix: render targeters after other overlays
After a lot of digging, it turns out that these are simply assembled in
the wrong order, so the targeter (ray) tiles are always drawn first. In
local tiles, the way alpha blending works means this doesn't matter (or
at least, I'm pretty sure that's what is going on), but in webtiles the
effect was that the targeter goes under other tiles. In particular, it
layered under wave tiles. This commit just brute-force pulls them out of
the overlay list and draws them last in this list. Possibly these
should be rendered even later.

Also, for future people digging into this: `draw_ray` does not draw
these ray tiles. I don't know what this function does draw.