fix: adjust targeter highlight colours
In terms of the practical effect, this ups the alpha values so that the
highlighters are more opaque; I also did a bit of color adjustment and
moved the dimmer highlight a bit towards gray. This makes them more
visible against blue and brown backgrounds, as in shoals. The new colors
are probably uglier, but they are definitely more visible. More tweaking
may be required. In the long run, I do wonder if some other color than
yellow might work better here, or even something more complex like
showing tile borders.

Under the hood, this replaces the targeter highlight tiles with
programmatically generated tiles, and fills out the tile tool with some
commands to make this possible to do.