fix: don't credit the player for wrath kills (Turukano)
Wrath attacks reuse shadow creature code, and because of this, wrath
kills were credited in the same way that shadow kills were credited.

A version of this was supposedly fixed in 9176aad88b6178, but I don't
think it actually was. The relevant check is not to the monster's actual
mindex; in mon-death.cc killer_index is simply reset to the player with
the actual mindex discarded as long as mons_is_player_shadow returns
true. (As noted in the comment there I think this would all work better
if the wrath code didn't reuse the player shadow code in entirety, this
is a recipe for bugs to sneak in.)

(cherry picked from commit f9c1391b642bdfc23306d39a96dc3ab46fbae2f5)