feat: a better wizmode interface to feature placing
Sometimes for debugging purposes it is helpful to set up specific
dungeon layouts in wizmode, but the interface via `&` for doing this was
really quite painful. This commit adds a much better interface for
placing a bunch of features quickly, via some wizmode-specific abilities
that provide a helpful targeter. It may or may not be a good idea to
have wizard-specific abilities (none exist prior to this), and if it
turns out to be too annoying, I can convert these back to something more
like the old `&` interface and selection. But for now, let's try this --
they are very convenient to access.

Some further features it would be nice to have if there's demand (but
since none of this is user-facing, low priority):
* a way to change the feature being built within the targeting
  interface, right now you need to use the separate ability.
* some way to create feature_property_types, right now you have to use
  the old UI to do this.
* a better feature selection interface, what's there is the old one
  split off into its own function.
* a way of toggling xray vision in the targeter? This is definitely most
  useful with xray vision on.

(Is this the beginning of dcss creative mode???)