fix: enforce range in blinkbolt targeting (kate)
Previously, you could use ! or @ to target an enemy out of range, and
blink to the end of range along the bolt path. This imposes the
targeting restriction from power leap and line pass that also requires
the monster target to be in range. (It also renames the relevant enum
value and clarifies the comments in the enum.)