fix: rework fungal movement behavior
Previously, fungus movement constraints were checked (at least) once per
square they consider moving to. For deathcaps and wandering mushrooms,
this is fine; their constraint is absolute. For lurking horrors, this
led to some odd behavior: the `random2` call is rerolled for every
possible target square, leading to a chance for movement that is a bit
hard to reason about, and determined largely by the degree of openness
around the monster. Various people have also reported that lurking
horror movement patterns have changed, with them becoming more likely to
charge and having weirder ("wiggly") paths. I *think* that to the extent
this is true it is because of subtle interactions with changes in
monster pathing over time: monsters consider more path options than they
used to (I think), and so the odds of them charging at a distance have
changed, and their paths are drawn from a wider variety of indirect
paths. There may also have been an interaction with the introduction of
squarelos. Though I have tested them with versions going back to 0.12
and the changes are not gigantic.

This commit moves the movement check out of the low-level movement code,
and so does the random check only once. This means that a lurking horror
always has exactly 1 in LOS+1 chances of moving at the edge of LOS, rather
than up to 5 in LOS+1 chances depending on the terrain.

One noticeable side effect of this change is that a wandering mushroom
at the edge of los will follow a foe as they move out of los, rather
than giving the foe a turn out of los. This is because the sequencing is
that the player moves, the fungus is out of los, and then can move. On
the old code, the movement target was also considered, perhaps for this
exact scenario. So if players find this too confusing, it may be
necessary to restore some version of it at least for wandering mushrooms
and deathcaps. I don't think lurking horrors would need it.