feat: New monsters, Walking Tomes and Living Spells
Courtesy of PleasingFungus. Walking Tomes are monsters that summon Living
Spells, which expend themselves to cast their spell unless they're destroyed
quickly. The tomes summon multiple spells at once, and the spell set chosen
is selected so that each one should require somewhat different tactics to deal
with. The current tome set is petrify, lcs, smiting, iceblast, but the theme
may be contracted or expanded as necessary. Tomes are intended to serve as a
component of the Tukima wizlab and may have potential as a lategame monster.