refactor: improve Menu action api a bit
Previously, the general on_single_selection function needed to handle
selection for every MenuEntry, leading to things like conditioning on
hotkeys in an awkward way, code being split up, etc. -- at least for the
Menu use cases I've been introducing in the 0.27 cycle (macro submenus
and the like). This commit lets a MenuEntry be defined directly with an
action in the form of an on_select function, making for more coherent
(though not less verbose) code in macro.cc.