refactor: move padded_bounds to a layout library
This function is used with the various omnigrid gridding functions to
put a bit of padding to allow the grid to expand into the edge of this
padding. There were two versions of it, one in layout_city the other in
layout_grids; layout_cellular was implicitly using one of them without
an explicit include.

This moves the function to omnigrid so that we have only one copy.