feat: shrink Hell subbranches on floors 1-6
This commit converts hells layouts to use the new dgn.builder_bounds()
instead of dgn.max_bounds(), adding parameterization on grid size for
some procedural values that were previously hard coded.

The dive portion of the hells can be exciting, but quickly becomes a
slog if the player must make the dive a second or third time, due to how
the stairs work. By shrinking the levels we reduce the time a player
spends with unfun repetition and increase the time spent in "good dive
moments". A subsequent commit will reduce the number of stairs spawned
on each level to 1 guaranteed stone stair.