Bring back Tukima's Studio
Removed in 0.17 for being a pushover, we have the technology (new monsters) to
bring it back now. The monster set is beefed up with rakshasas and dancing
launchers as well as the brand new walking tome enemies and reworked strange
machines, which will summon shadow creatures and Mara style player illusions.