fix: try to place the entire requested number of maps with place_maps
Since 2010, the dgn.place_maps function would attempt to place the
requested number of maps... until one failed to place for any reason,
then it would give up. Before then it would error, so this was an
improvement. However the commit messages around that time suggest that
the goal was for the function to continue in its attempt, and return the
total number of placed vaults.

This change should hopefully make some of our serial vaults shine a bit
better. If they become too overwrought we might need to retune the
number of serial maps each one places, but in my local testing it seemed
ok.