feat: new unrand artefact: dreamshard necklace
The dreamshard necklace is an amulet of nothing that for the most part
does nothing. However, on a lethal or near-leathal blow the necklace
shatters, absorbing the blow and releasing protective dreams of healing
(a big heal, between half and full HP), friendship (shadow creatures),
and clouds (fog).

That is, the player gets to trade their amulet slot (with its many
powerful effects) for a one-time-get-out-of-death-free card.

[ Committer's note: Simplified the shattering effect significantly,
  re-factored the trigger logic to be simpler, moved the dreamshard
  effect to spl-goditem, reduced the visuals to a single flash (no
  permanent tile changes), implemented Ashenzari and Death's Door
  checks. Thanks to kate and PleasingFungus for input in the PR process.
  Closes #2078. ]