fix: don't use homebrew includes over contribs
Mac clang imposes `-I /usr/local/include` on builds, which takes
precedence over our use of `-isystem` and leads to an unfortunate
interaction with homebrew. Regular mac build tools don't actually use
this location, and nor does macports, but homebrew does. For example if
the user has SDL2 installed via homebrew, those includes will be used
even though the build tries to link with the contrib version of SDL2,
resulting in build failure. (This has been reported for a couple of
years, it seems.)

There may be a number of ways to fix this, and I'm not sure I'm a fan of
using -isystem for contrib libraries that we are providing and
maintaining and that aren't located in a system include path, but this
commit takes a more minimal approach to working around this issue, just
ensuring that checking this directory takes precedence under any
`-isystem` includes introduced in the Makefile. (Of course, it's
arguably a bug that mac clang does this with `-I` instead of `-isystem`,
but good luck getting that changed.)  Possibly in the future `-isystem`
should be switched back to `-I` for contribs so that we have better
control over this (by using `-isystem` we ceded control to any way of
setting `-I` e.g. via environment variables), but doing so would require
a lot of testing.  (See e84909d808a7 for where this originated; from the
dev logs I suspect this may be partly about suppressing warnings from
include headers as well.)