fix: don't crash on fedhas wave wrath
Via side effects, this wrath action results in `move_player_to_grid`
which puts the player on the same square as the fake monster casting the
spell. This monster reuses shadow creature code, but as of f9c1391b6 no
longer returns true for mons_is_player_shadow, so an ASSERT got tripped.
This commit just adds another similar function for god wrath so that at
least this special case can be checked without knowing implementation
details, but also this implementation of wrath is not ideal.