fix: refix something that is *not* an off-by-one-error
da2e87d06967ae turns out not to have been an off-by-one error, but
rather the result of bad initial calculations when the menu is not yet
displayed, resulting in invalid scroll values. In addition,
da2e87d06967ae broke scrolling in menus that don't use a hover. This
wasn't noticed before now partly because webtiles still works, the bug
seems to be a bit erratic, and in the words of one dev, "oh so that key
is actually broke and its not my broken keyboard".