Adjust some stat zero effects
Replace 0 strength's regeneration reduction with -Regen in monster LOS (the
same as the mutation), to give it an effect in combat without making
resting take excessive time. Also replace 0 intelligence's 4/5 chance for
scroll-reading to fail with -Scroll (since it no longer needs to be
possible to read remove curse scrolls to unequip cursed -Int gear).

This whole mechanic could probably be simplified further since it's such a
rare occurence.