Make the robe of Clouds' thunder passive
So as to continue moving away from activated abilities on equipment (as
with other unrands such as the cloak of the Thief and ratskin cloak),
give the robe of Clouds a chance of creating a storm cloud on enemies
within radius 2 each turn. Also mention its cloud immunity in the
description.