refactor: handle self-targeting checks better in beam.cc
Xom's chessboard fakes smite targeting for a variety of effects by just
setting the beam source to a target monster's position. This leads to
some conflation of real self-targeting and fake smite-targeting in a
bunch of subtle corner cases, and also led to 12637. Rather then handle
this on case-by-case basis as in 395adb26f73b for poly and a708605c5b
for teleport, this commit restores the general check from
pre-395adb26f73 but makes it more restrictive. It also sets the agent
for xom's chessboard actions (as the player, tbc) in the course of this.
This commit should also fix some related weirdness with wrath that
hasn't yet come up.