Reduce most summon spell caps to 1
Summoning magic has long had the downside of kills by allies granting only
50% XP, which works fine in terms of balance due to how hugely powerful
summoned allies are, but is very inelegant as well as just feeling bad to
many players.

The main problem caused by summon spells is how easy it is to quickly and
repeatably create a very large number of allies, so to try and balance this
better and allow removal of the XP penalty, greatly reduce the summon caps
for player-castable summoning spells. Almost all spells are capped at 1
summon, with a few exceptions for spells that wouldn't function with such a
low cap (such as Summon Small Mammal where the individual allies are very
weak), for a few high-level spells (Summon Hydra and Monstrous Menagerie)
which are instead capped at 2, and for the very highest level spells (such
as Summon Horrible Things and Dragon's Call) where the high level of
investment required can justify their strength.

Being able to create allies on demand is still fundamentally an extremely
powerful ability so this is very unlikely to cause problems for summon-
focused characters - it'll just be required to either hybridise more or to
use a wider variety of summon spells.