Change the spelling of some words to Australian English.
Change a few places where words were spelt with -eling at the end and
where this is not a correct spelling according to
https://www.australian-dictionary.com.au/wordcheck/.

This only changes things in places where the spellings may be displayed
to the user. I include the documentation in this. I haven't changed the
changelog, or any of the contrib stuff.

This changes:

channeling -> channelling
dueling -> duelling
groveling -> grovelling
quarreling -> quarrelling
raveling -> ravelling
traveling -> travelling
tunneling -> tunnelling

Closes #2093