Make Shadow Creatures a monster-only spell
It doesn't work well with the reduced player summon caps, since a big part
of its functionality is being able to create monster bands, and the
interaction between bands and the summon cap was already fairly messy.
Being able to summon rare branch-specific monsters and big monster bands is
a fun effect, but works much better on the already-existent scroll of
summoning, which avoids the problems of summon caps and is balanced by
being a consumable.