feat: script options for checking vault opacity
This new set of options to the placement checker script is designed to
find vaults that are not marked as transparent but probably should be.
To use this, make sure to test with enough iterations so that
randomization can play out. It's actually also somewhat useful for
finding vaults that aren't marked as transparent and have a teleport
island, when going through vaults one by one.