fix: transparent tags for some vaults
It's better to mark vaults transparent than not, because that allows the
builder to check for teleport closets itself, and also allows the
placement script to more easily check that the intention indicated by a
tag matches with what happens when the vault places. This commit has a
batch of fixes for vaults that automated testing indicates should be
marked transparent. This stops in a somewhat arbitrary place as I was
working on this several months ago and haven't managed to get back to
it, but eventually this should be checked for all vaults.

This commit also includes a few teleport closet fixes for vaults that
didn't immediately show up as being marked transparent.