fix: prevent bad array access for describing unid'd staves
Also add a bunch of other checks around this particular crash; the basic
problem here was that SK_NONE is greater than NUM_SKILLS and a lot of
code that has skill-related arrays doesn't check bounds and stops at
NUM_SKILLS. The specific crash was because `staff_skill` returns SK_NONE
in cases where it can't find a matching skill, which happens when a
staff is unid'd.  Because the bug depended on what was in adjacent
memory, the crash was somewhat variable.