vaults: Dis end vault monster set adjustments
In line with the Dis changes, remove trash mobs and replace with the new
Dis cast. As with the non-end vaults, a few juggernauts sneak in as a
guest threat, and the cacodemons in the one grunt ending stick around.

Might have been too heavy with caustic shrikes on some of these.