Shrink some overview screen labels to make them fit better
Reduce the width of the first column of labels from 9 to 8. This creates a 1 character gap between the right edge of the Stlth field and the second column of labels.

Set the length of the Stlth, HPRegen and MPRegen labels using cwidth. This means that every field in the two label columns has its width set by cwidth.

Remove loops from _stealth_bar().