Disable Okawaru Duel in Sprint
Instead of crashing, disable the ability in Sprint, since it's not supposed
to be possible to leave the single D floor by any means (and the Abyss is
also disabled). While it'd be possible to special-case the Arena and enable
the branch, it seems likely to break Sprint balance even more than usuali
(especially in maps like linesprint and arenasprint, for example).