Have C-glyph altars use temple_god_list()
Previously there was a hardcoded list of gods that couldn't generate for
C glyphs in dungeons. It turns out this was essentially equivalent to
the list already used in temple_god_list(), so this removes that
duplication. This also fixes an issue where an altar to Ignis could
generate on a few Zot dungeon levels which used random altars.