refactor: convert staff of dispater evoke to an ability
This allows for better transparency for the costs, more code reuse, etc.
I have provisionally converted the HP cost to a percentage, which
overall should work better for damnation. This is currently set at 10%,
and it may be a nerf to many characters at higher HPs. Further
adjustment might be required.