refactor: converk staff of olgreb evoke to an ability
As with the previous two commits, this lets a much more straightforward
version of the power be shown to the player. As part of this, I have
simply removed the fail check. This required 5 evo skill to get to 100%,
which is trivial; it originates from earlier versions of the staff that
could also cast venom bolt.