feat: Despoiler loot and monsters in a Lair ending
For due_jungle_book, the thematic loot item guarded by the obsidian
statue is at a position where one had to step in view of the statue to
see the item at all. Additionally the statue is easy to autoexplore
into, and the area in which it resides is rather cramped for it to be
able to use its summons.

This commit encloses the loot item in a clear stone room with a clear
door, placing the statue in front. The room is visible from another room
usually encountered before the statue area. Hence the player can always
see basic information about the loot item before deciding to fight the
statue, and the statue itself will likely be seen and autoexcluded
before it's reached.

An additional bonus is that the loot item can't simply be apported
without fighting the statue, since the door must be opened first.
Finally, we clean up the terrain surrounding the statue to open up space
a little, allowing its summons to work better.