feat: downgrade Fighters to a buckler
The starting kite shield on a Fighter was in a weird place. It had the
kind of very long term value that the recent changes to spellcasting
backgrounds have moved backgrounds further away from. It was completely
reasonable to put some enchant armour scrolls into it and use it
throughout the game. Unlike the aux armour of other backgrounds which
provide 0-1 AC at the start and 2-4 AC at best, the kite shield starts
out around 8SH and without enchants or increased dexterity rises to 13SH
with training; it is a significant portion of the user's defenses.

On the other hand, with the 0.27 change linearizing shield delay
penalty, the shield was a liability for normal sized races. The penalty
was sufficiently high that "remove your shield turn 1" advice
reappeared and was not unreasonable. This is unfortunate, as a starting
kit should ideally be useful at game start.

Given the above, lets replace the kite shield with a buckler. The
buckler will be worth using on turn 1, but any normal-sized heavy melee
fighter will want to find an upgrade... or consider a two-hander or an
alternate character archetype, depending on what drops. Races that
cannot use a buckler get a kite shield.

Currently this leaves the background description and manual intry as
is---both simply mention "shield".