fix: adjust hell effect strength
The timing reduction in the final balance got hell effects in about the
right place in terms of frequency, and contam in the right place for
quantity.

Drain was checking rN, which players carry a lot of. Instead
of turning it up to big number, this commit makes hell effect drain
ignore rN.

Stat drain was hitting with reasonable frequency, but the XP
available made it worked off easily. This commit increases the amount of
stat drained.