Allow / to be used instead of shift in UIDirectionChooserView.
/+direction and shift+direction are valid combinations for the "long walk" command. Shift+direction is a valid combination for "fire in a direction", with direction+! as an alternative, but having /+direction do the same should make it easier to remember.

Put ' and ; before * and / in targeting_help_1, as the above change means the latter keys are used as modifiers if use_modifier_prefix_keys is set (the default).

In unmangle_direction_keys(), only let * modify characters which CONTROL() maps to \x00-\x1F. In particular, this means that ** is *, giving a work-around for people who use the default settings but are used to * and / cycling through objects.