fix: improve `auto_switch` behavior with two bad weapons
This makes the check on the current weapon and the target weapon
identical, following a suggestion in #2057. I was initially worried that
this would lead to further special cases but I think it's ok.

This commit also makes an item definedness check explicit, where before
it relied on `is_melee_weapon` to do the right thing.

Resolves #2057