fix: don't corrupt place_info on &ctrl-r in the vestibule
As usual, place_info handling is a complete mess and extremely brittle.
This crash could happen because the vestibule is a depth 1 connected
branch, so a fixup corrected levels_seen in files.cc before getting to
the wizmode handling for this field for this command, which assumed that
it was uncorrected. This commit adds a workaround on top of the fixup.
The fixup was originally created for a very specific bug but is *also*
necessary for `&ctrl-r` in the vestibule, because otherwise the
following assert crashes due to the branch depth. Possibly place_info
validity handling should be entirely rethought.

Resolves #2067