build: improve messaging if `git describe` fails
This message was pretty cryptic and didn't cover one major case at all
(when `git describe --tags` fails because the repository is a shallow
clone). The new error is extremely verbose but hopefully provides much
better hints as to what went wrong.

This also adds a comment documenting a workround to gen_ver.pl. This
workaround seemed too niche to me to put in INSTALL.md, but maybe it
should be there?

See discussion in #2115