Rework the flesh golem
This Cigotuvi wizlab monster is a reskinned iron golem, so it got buffed when
they did. While a wizlab-depth player can certainly deal with a slow 300 hp
monster, it feels pretty awful to have to do it. Lower it to ~100 hp and give
it corrupting pulse so it has something to do besides walk slowly at you.