Fix: don't list walljump twice while silenced
This specifically affected the "Abilities" section of the sidebar, but
I'm not sure where else. It seems likely that this is legacy code from
when walljump was more of a pseudo-ability, but since whether something
can be used while silenced is determined by the failure basis (e.g., if
it's an invocation), there's no need for this check to exist in
get_god_abilities anymore.

Closes #2120.