Correct logic for walljump in the ability menu
The ability menu specifically calls get_god_abilities with
ignore_silence set to false, which is different from the way the tiles
buttons call it. This means we have to also respect that flag, to ensure
walljump shows up in the ability menu when it should.

This should (and probably will) be generalized into an actual ability
flag that marks an ability as ignoring silence.