Fixes two unrelated bugs concerning Phantom Mirror and the monster version of Phantom mirror.
Bug 1 is simple: When a monster casts phantom mirror on another monster,
the casting monster is unentangled/unconstricted. This behaviour is unintended,
as clearly, the monster being copied should be freed instead.

Bug 2 is more subtle: When a monster is copied, its constriction status is
not copied (as intended). However, the enchantment applied by BVC to indicate
direct constriction IS copied. This results (in tiles) in a graphical bug, where
the monster has the BVC icon, but can move around and act as normal.