Make a BIG removal
Remove SIZE_BIG, switching monsters into either SIZE_LARGE or SIZE_GIANT
as seems appropriate.

Following up on 7bd0902ce6cc84c6, we really don't need two sizes that
are 'bigger than any player race'. Is a yaktaur bigger than a centaur?
How does a spatial maelstrom compare in size to a fire dragon? These are
questions that we don't care about and shouldn't bother with.

This is a small buff to dragon form and maybe a nerf to something else.
Who knows.