fix: step up lua throttle aggressiveness
On testing, a tight loop in ready() takes nearly 3 minutes to get killed
by this on a fast cpu, and does still have a noticeable cpu impact until
then -- I expect it'd be much worse on a busy server. This commit makes
the throttle kill scripts much more aggressively, and also, if ready()
in particular is killed by the throttle, keeps it from running again. If
this runs up against legitimate online use we can adjust, but I really
think it won't. I tested this on qw, and qw does trip the throttle but
never more than ~4 times in a row at the absolute max. So 15 is still
extremely conservative. It's possible that this could be tightened even
further: e.g. by starting the throttle time at something like 128ms
instead of 2, or by reducing the number of lua instructions this is
checked between (currently defaults to 50000 I think).

No impact outside of webtiles.