fix: don't allow c_message reentry
This hook is already pretty wild, given the arbitrary points where mpr
can be called from, but it's additionally very, very easy to trigger
recursion that would be infinite if it weren't for the lua stack limit.
Depending on what happens this recursion can be quite cpu intense as
well, even with the stack limit. This commit just prevents any recursive
triggering of the hook in a heavy-handed way.

(Overall, I find this hook quite worrying, and I suspect it still needs
substantial auditing.)