fix: improve SIGHUP handling in clua code on dgl builds
User lua that involved hand-rolled prompts could interact in a bad way
with SIGHUPs, prevent the game from ever properly handling the HUP.  The
following example based on long-time crash generator `ziggurnaut` is a
case in point:

    {
    function ready()
        force_prompt("asdf")
    end

    function force_prompt(text)
      while true
      do
        crawl.formatted_mpr(text)
        crawl.formatted_mpr("Type p", "prompt")
        res = crawl.getch()
        if string.lower(string.char(res)) == "p" then
          break
        end
      end
    end
    }

If the player disconnects at the prompt, the `while` loop runs in a
tight loop because getch returns immediately without a client, and the
loop never exits on its own, so the HUP isn't handled. The webserver
does detect this 10s or so later, and kills the process with SIGABRT, so
the issue isn't catastrophic, but it still is 100% cpu for that time
period and generates an annoying crash report in #crawl-dev. (For
reference, the best thing to do here on the current API would be to also
accept ESC (27) at any prompt implemented in a loop like this; getch
returns ESC on disconnects. Or just use crawl.yesno instead of trying to
write a custom one.)

This commit piggybacks on the clua throttle code to check for HUPs if
the throttle trips; if a HUP has been seen it immediately errors out of
clua code, which leads to a graceful shutdown of the process. It's
possible this could use its own separate hook (which would let it work
by default on non-DGL builds as well, and possibly be checked more
often), but the main problem here is on webtiles so I'm going with this
more minimal solution for now.

(cherry picked from commit 34660d742963940429c2167909249a3764e99b35)